+++ m4lcommon/include/kopano/RecurrenceState.h
@@ -69,7 +69,7 @@ namespace KC {
 class _kc_export RecurrenceState _kc_final {
 	public:
 	HRESULT ParseBlob(const char *lpData, unsigned int ulLen, ULONG ulFlags);
-	HRESULT GetBlob(char **lpData, unsigned int *lpulLen, void *base = NULL);
+	HRESULT GetBlob(char **lpData, size_t *lpulLen, void *base = NULL);
 
 	class _kc_hidden Exception _kc_final {
 		public:
