+++ swig/RecurrenceState.swig.h
@@ -100,7 +100,7 @@ public:
 class RecurrenceState {
 public:
 	HRESULT ParseBlob(const char *lpData, unsigned int ulLen, ULONG ulFlags);
-    HRESULT GetBlob(char **lpData, unsigned int *lpulLen, void *base = NULL);
+	HRESULT GetBlob(char **lpData, size_t *lpulLen, void *base = NULL);
 
 	/* not inlined for swig */
